Original version of code.
Currently, it allows user to load a G-code file. The program uses the
cuts described in the G-code to recreate the original drawing, which is
displayed in a image box. It then allows the user to add text of varying
sizes on the drawing, which are saved into an updated G-Code.